repository is building

+++ Services/Common/IM.Service.Common.Net/Models/Dto/Http/Companies/StockVolumeGetDto.cs
@@ -7,6 +7,6 @@ namespace IM.Service.Common.Net.Models.Dto.Http.Companies
         public string Company { get; init; } = null!;
         public DateTime Date { get; init; }
         public string SourceType { get; init; } = null!;
-        public int Value { get; init; }
+        public long Value { get; init; }
     }
 }
